move change-timer to editable
